+++ iExtraExample/ViewController.swift
@@ -28,10 +28,8 @@ class ViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        let view = UIView(frame: CGRect.init(x: 100, y: 100, width: 100, height: 100))
+        let view = PannableView(frame: CGRect.init(x: 100, y: 100, width: 100, height: 100))
         view.backgroundColor = .red
         self.view.addSubview(view)
-        view.startRotation(withDuration: 10)
-        view.applyCornerRadius(Corner())
     }
 }